small fixes

+++ src/main/java/com/acmebank/repository/AccountRepository.java
@@ -4,7 +4,7 @@ import com.acmebank.domain.Account;
 import org.springframework.data.repository.CrudRepository;
 
 public interface AccountRepository
-    extends CrudRepository<Account, Long> {
+    extends CrudRepository<Account, Double> {
 
   Account findByAccountNumber(int accountNumber);
